Work around a strange bug in Firefox and WebKit that causes click events to be fired even though the user dragged the mouse.
Deal with this by:

- Cancel dragstart instead of mousemove.  This still prevents dragging,
but is only triggered if the browser actually would have considered it
a drag, eg. if the mouse has moved more than a couple pixels.  This is
the point where clicks should no longer happen.

Record when this happens.

- If we've cancelled a drag when the release happens, cancel the upcoming
click (if any).  The click event happens after the mouseup, so we need to
set an event listener and then remove it from the event loop.  This is
done on the window from a capturing event handler, so delegate event
handlers don't receive the event either.

It's important that we don't just set this.dragged from mousemove, since
that would prevent any thresholding of clicks at all.